iaso_extract_metadata/requirements.txt
XlsxWriter
psycopg2-binary
openhexa.toolbox==2.7.0